Add a warning to task_failure
It may be clear to others but it sure wasn't for me, that task_failure will not be dispatched when task_always_eager is True. I am not sure if this patch is the correct way to document this and I am happy to modify it if needed. However, documenting it somewhere might help the next person.